Separated code out into API files
